form/pyproject.toml
[project]
name = "form"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "sigvardbratlie", email = "sigvard.bratlie@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.47.0",
    "google-cloud-bigquery>=3.40.0",
    "streamlit>=1.52.2",
]

[project.scripts]
form = "form:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"
